core-api: move utility api system implementation into apis/system
