+++ packages/core-api/src/apis/definitions/AlertApi.ts
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { createApiRef } from '../ApiRef';
+import { createApiRef } from '../system';
 import { Observable } from '../../types';
 
 export type AlertMessage = {
