+++ packages/core-api/src/apis/definitions/ConfigApi.ts
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { createApiRef } from '../ApiRef';
+import { createApiRef } from '../system';
 import { Config } from '@backstage/config';
 
 // Using interface to make the ConfigApi name show up in docs
