+++ packages/core-api/src/apis/definitions/StorageApi.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { createApiRef } from '../ApiRef';
+import { createApiRef } from '../system';
 import { Observable } from '../../types';
 import { ErrorApi } from './ErrorApi';
